pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "carl-studio"
version = "0.3.0"
description = "CARL (Coherence-Aware Reinforcement Learning) — information-theoretic reward signals for LLM training via token-level probability distributions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Tej Desai", email = "tej@terminals.tech" },
]
keywords = [
    "rl",
    "reinforcement-learning",
    "coherence",
    "grpo",
    "training",
    "llm",
    "information-theory",
    "reward-shaping",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Operating System :: OS Independent",
    "Environment :: GPU :: NVIDIA CUDA",
]

dependencies = [
    "pydantic>=2.9",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "rich>=14.0",
    "numpy",
    "anyio>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
training = [
    "torch>=2.4",
    "transformers>=5.0",
    "trl>=0.12",
    "peft>=0.13",
    "datasets>=3.0",
    "bitsandbytes",
    "trackio",
]
observe = ["anthropic>=0.40"]
hf = ["huggingface-hub>=0.25"]
runpod = ["runpod>=1.7"]
tinker = ["tinker>=0.16"]
tui = ["textual>=1.0", "textual-plotext>=0.3"]
mcp = ["mcp>=1.0"]
dev = ["pytest", "ruff", "pyright"]
all = ["carl-studio[training,observe,hf,runpod,tinker,tui,mcp,dev]"]

[project.scripts]
carl = "carl_studio.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/carl_studio"]

[tool.pyright]
include = ["src"]
typeCheckingMode = "strict"
pythonVersion = "3.11"

[tool.ruff]
target-version = "py311"
line-length = 100
